refactor: Вынести код работы с базой данных в пакет DAO
