+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import ru.yandex.practicum.filmorate.dao.MpaDao;
 import ru.yandex.practicum.filmorate.model.Mpa;
-import ru.yandex.practicum.filmorate.storage.Impl.MpaDbStorage;
 import java.util.List;
 
 @Slf4j
@@ -15,17 +15,17 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/mpa")
 public class MpaController {
-    private final MpaDbStorage mpaDbStorage;
+    private final MpaDao mpaDao;
 
     @GetMapping
     public List<Mpa> getAllMpa() {
         log.info("Request to get all mpa");
-        return  mpaDbStorage.getAllMpa();
+        return  mpaDao.getAllMpa();
     }
 
     @GetMapping("/{id}")
     public Mpa getMpaById(@PathVariable Integer id) {
         log.info("Request to get mpa by id");
-        return mpaDbStorage.getMpaById(id);
+        return mpaDao.getMpaById(id);
     }
 }
