+++ src/main/java/ru/yandex/practicum/filmorate/dao/GenreDao.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.filmorate.dao;
+
+import ru.yandex.practicum.filmorate.model.Genre;
+import java.util.List;
+import java.util.Optional;
+
+public interface GenreDao {
+    Optional<Genre> getGenreById(Integer id);
+
+    Optional<List<Genre>> getAllGenres();
+}
