+++ src/main/java/ru/yandex/practicum/filmorate/dao/Impl/GenreDaoImpl.java
@@ -1,9 +1,10 @@
-package ru.yandex.practicum.filmorate.storage.Impl;
+package ru.yandex.practicum.filmorate.dao.Impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.controller.GenreController;
+import ru.yandex.practicum.filmorate.dao.GenreDao;
 import ru.yandex.practicum.filmorate.exception.ObjectNotFoundException;
 import ru.yandex.practicum.filmorate.model.Genre;
 import java.sql.ResultSet;
@@ -15,11 +16,12 @@ import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
-public class GenreDbStorage {
+public class GenreDaoImpl implements GenreDao {
     private final JdbcTemplate jdbcTemplate;
     private static final String SELECT_GENRE_BY_ID = "select * from genres where genre_id = ?";
     private static final String SELECT_ALL_GENRES = "select * from genres order by genre_id";
 
+    @Override
     public Optional<Genre> getGenreById(Integer id) {
         return Optional.of(Objects.requireNonNull(jdbcTemplate.queryForObject(SELECT_GENRE_BY_ID,
                                 (rs, rowNum) -> makeGenre(rs), id)).stream().findFirst())
@@ -27,6 +29,7 @@ public class GenreDbStorage {
                         GenreController.class.getSimpleName()));
     }
 
+    @Override
     public Optional<List<Genre>> getAllGenres() {
         return Optional.of(jdbcTemplate.queryForStream(SELECT_ALL_GENRES, (rs, rowNum) -> makeGenre(rs)).findFirst()
                 .orElseGet(ArrayList::new));
