+++ src/main/java/ru/yandex/practicum/filmorate/dao/Impl/MpaDaoImpl.java
@@ -1,9 +1,10 @@
-package ru.yandex.practicum.filmorate.storage.Impl;
+package ru.yandex.practicum.filmorate.dao.Impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.annotation.Validated;
+import ru.yandex.practicum.filmorate.dao.MpaDao;
 import ru.yandex.practicum.filmorate.model.Mpa;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -12,15 +13,17 @@ import java.util.List;
 @Component
 @Validated
 @RequiredArgsConstructor
-public class MpaDbStorage {
+public class MpaDaoImpl implements MpaDao {
     private final JdbcTemplate jdbcTemplate;
     private static final String SELECT_MPA_RATING_BY_ID = "select * from mpa_ratings where mpa_id = ?";
     private static final String SELECT_ALL_MPA_RATINGS = "select * from mpa_ratings order by mpa_id";
 
+    @Override
     public Mpa getMpaById(Integer id) {
         return jdbcTemplate.queryForObject(SELECT_MPA_RATING_BY_ID, (rs, rowNum) -> makeMpa(rs), id);
     }
 
+    @Override
     public List<Mpa> getAllMpa() {
         return jdbcTemplate.query(SELECT_ALL_MPA_RATINGS, (rs, rowNum) -> makeMpa(rs));
     }
