+++ src/main/java/ru/yandex/practicum/filmorate/dao/MpaDao.java
@@ -1,0 +1,10 @@
+package ru.yandex.practicum.filmorate.dao;
+
+import ru.yandex.practicum.filmorate.model.Mpa;
+import java.util.List;
+
+public interface MpaDao {
+    Mpa getMpaById(Integer id);
+
+    List<Mpa> getAllMpa();
+}
