+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -2,15 +2,13 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import javax.validation.constraints.NotNull;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class Genre {
-    public Genre() {
-
-    }
-
     @NotNull
     private Integer id;
     @NotNull
